refactor(taskfiles): rename unlock-local to unlock
Renamed the `unlock-local` task to `unlock` in Taskfile.yaml to align with the existing `unlock` task functionality. Updated the description to reflect that it unlocks a restic source repo from a local machine. Removed the old `unlock` task that was intended to unlock all restic source repos, as it is no longer needed. Adjusted the unlock.yaml.j2 template to remove the restic unlock command that was part of the old `unlock` task.